Reorder tone mark and nikhahit in DroidSansThai.ttf
Fix for bug 6630331.

This is a workaround for a shortcoming in the Thai shaper in Harfbuzz.
De facto (but not documented) shaping rules dictate that a tone mark
followed by a SARA AM should be rendered with the NIKHAHIT below the
tone mark, rather than above it as would be indicated by a
straightforward interpretation of the GSUB and GPOS rules. This patch
adds an additional substitution of the form MAI THO + SARA AM ->
NIKHAHIT + MAI THO + SARA AA. The three tone marks affected are MAI EK,
MAI THO, and MAI CHATTAWA, based on an usage analysis of the Thai
dictioary included in ICU.

When we upgrade to harfbuzz-ng, we can revert this workaround.

Change-Id: I209ed094e7c89ea7f6fd6e983d37e396c6da1f64

Former-commit-id: f408bf847a250f2d914a50de151cd63650601be5