Fix requestLayout-during-layout issue with Gallery
The way that the Gallery was adding views (which it does in layout)
triggered a requestLayout() because of a call to setLayoutParams()
in View.addViewInner(). This tripped up the new logic that makes
layout-in-layout work, causing layout to be called on every frame,
even when the app was not doing anything.

The fix is to call the appropriate version
of requestLayoutInner() to tell it to skip calls to requestLayout().

Issue #7714495 requestLayout in endless loop

Change-Id: Icdcd886d30e4dd972da36bc6e50732c7213530c5

Former-commit-id: bfea76c7879306f62338546665e8d347197ea179