Fix re-enabled state of up after action view collapse
A certain call ordering to the action bar could cause the record of
the pre-action view expanded home button enabled state to be
incorrect. Store this in all cases, not just when the action view is
initially expanded. (Except for when we actually disable it for the
expanded view.)

If an action view is already expanded when we get a call to change the
enabled state of the home/up button, don't modify the state of the up
container or other accessibility info, but record the new state if
requested.

Bug 8142097

Change-Id: I070c151aa9f2046d2114ee541d964bf68c6f0def

Former-commit-id: 07cf971e58aecab76f2b5b575cfe62120005c6e4