Fix for bug: 8184401 - GridLayout accounts for child's size/measurements even when a child is "GONE"
The original fix:

https://googleplex-android-review.googlesource.com/#/c/273122/1

was incorrect as it left GridLayout to simply ignore views with GONE status.
Instead, views with GONE status must cause their corresponding rows and
columns to collapse to zero width/height in the case where the rows/columns
are not flexible.

The actual bug is the result of the fact that getBaseLine() is not parameterized
by the height/size of the view for which the baseline is being requested. In the case of
a view that is GONE, GridLayout will request the baseline of a component that has zero height,
and get results other than 0 (which is doesn't make sense). Prevent this with a specific check.

This CL should fix 8184401 without causing:

https://b.corp.google.com/issue?id=8192858

Change-Id: Ic74888b499466bbfac5fb97d49eede5b6c343127

Former-commit-id: a841644367772f51d4ac5f2af2a14cc2d0e36df1