Fix long-standing bug with LockPatternView drawing artifacts
Moving very slowly in LockPatternView would often draw discontinuous
lines and have missing wedges in the pattern rings.

This fixes the problem by correctly accounting for the invalidate
region. The invalidate region is constrained to the last registered
pattern cell and the current point.  When a new pattern cell is added,
the bounding region expands to refresh it once.

Fixes bug 8315830

Change-Id: I72342535ae292fcadb02d96b89665ba8431886b2

Former-commit-id: 0caa377f4688f175ae22229a10294468610a116e