[DO NOT MERGE] DatePickerDialog title initially incorrect if calendar view shown.
If the calendar view of the date picker dialog is shown we
do not use a formatted date as the dialog title since the
calendar view shows it, instead we show a static "Set date"
title. However, the initial state of the date picker dialog
on a tablet has a formatted date since the date picker widget
does not do a correct check whether the calendar view is shown
which returns a false negative resulting in a title with a
formatted date being added since the dialog thinks it has no
calendar view.

Change-Id: Ibf4c433b34ab1ee1bc1dc386689664a32283fb6a

Former-commit-id: e10837f444c3da40a50a269e696afac63729bce3