Incorrect unbinding of accessibility services.
If no accessibility services are enabled, we disable the
accessibility event firing to save resources. When the last
such services is disabled the system was not unbinding. As
a result the user was seeing the touch exploration enable
dialog when the service that requested it is disabled. Also
there is one service the system is bound to that is not used.

bug:8439191

Change-Id: I6f37f2573a815bfb29870298aa0abbb1fa105588

Former-commit-id: cefac5acc845bc0fd36cb2c7e9e0992ecda33230