Reenable optional permissions parsing and expose APIs
Modify the package parsing code to understand optional permissions
(android:required="false"). This is essentially a rollback of
f6cc410422600897113d8d77ab07c8a72b31c1e1 [formerly e824120016248bcfe3962883fe9f3b0481dd1c3d] with some minor changes.

Expose the requestPermission API to third party apps. This allows
an app to request an Intent which, when passed to
startActivityForResult, will prompt the user to approve permissions
for an app.

In the event we decide to not launch with this feature, this change
can be rolled back.

Change-Id: Ie3626deae9b16e510323f94a2c80377f7c84b26f

Former-commit-id: 828c16f140eadb9265f2bf53469902dc7cdb2284