Switch indic fonts to Noto
This change switches from the older (cibu and lohit) Indic fonts to
the new Noto fonts. All the new fonts come in both normal and bold
weights, and with system ui (more compact vertical metrics, to match
Roboto) variants as well as normal.

The fonts have been tested on a phone (Galaxy Nexus) and seem to work,
but should be more carefully evaluated by native speakers.

Change-Id: Ib0a10ee5a130bab258ff198e22b8304e00141073

Former-commit-id: 6ff0bb6f5cc2678ee8f23e1c71faf1409af5df98