rework GLSurfaceView so it can use the choreographer
GLSurfaceView will now automatically use the Choreographer to
schedule render updates on VSYNC. This happens only in
RENDERMODE_CONTINUOUSLY.

GLSurfaceView was rewriten to use a Looper and a HandlerThread
instead of a custom thread.

The basic mode of operation is unchanged, however the implementation
is largely different.

Removed support for older GPUs which had a limited number of contexts.

Bug: 8485730

Change-Id: I7b270f7ad7bb1259e858429054b1c8c9cc159dc6

Former-commit-id: 57d019e222abc0de3f8876e682617d42872230c6