Uphold common ordering expectations around action bar Home/Up dispatch
Some apps aren't particularly happy if a stray key event is dispatched
to a newly created window before its menu is prepared, causing the
action bar's Home/Up event to be dispatched. Ignore these clicks on
the Home/Up button before the menu goes through its initial prepare
step.

Note that it is still possible (and valid!) for Home/Up to be
dispatched even if the app chose to return false from
onCreateOptionsMenu or similar.

Bug 7085070

Change-Id: If4b7d5f8c5a08ce8a094f1919347604d78ddedfb

Former-commit-id: 5ecacd667cc4a047ef9a4d89d218059230fa26e7