Fix for bug 8574856 - bad metrics in Roboto Italic
The last update to Roboto Italic has a stray point in delta.alt with a
y value of -11299. This causes metrics to give too much vertical space
in some cases.

This is a "hot patch" fix, with the stray point deleted using TTX. We
want to fix the font upstream as well, but this should fix things in
the mean time.

Change-Id: I08edae96f69753b2d46dfc3fd1cc3b9176dcb5ad

Former-commit-id: 98082fa98bb1dc62d02b1c7dfb000a7c412a1188