Fix for bug 8500202: Right-to-left text not showing in list view
In single line mode, changing the text from LTR to RTL (or vice versa)
affects the alignment, which in turn means that bringTextIntoView is
needed to update the scrolling. A registerForPredraw should be done to
make this happen, but it was missing. This patch tests explicitly for
direction changes in this case, and schedules a predraw if so.

Change-Id: I16e0e23141c244dc8adc00ea8306dfe4c9bf487d

Former-commit-id: f5cf6c9d4472e99caf4c9a3525d7b10cdf1e1c9f