BinderDied may be called after we unregistered the death recipient.
It is possible that an accessibility service's package was force stopped
during whose handling the death recipient is unlinked and still get a call
on binderDied since the call was made before we unlink but was waiting on
the lock we held during the force stop handling. Added a check whether the
service is already disconnected and if so do nothing.

bug:8600388

Change-Id: I4a9ca305b9863d986b930a7c1ec8f9006b16a333

Former-commit-id: 2602a3974aafc9e9be968e7bfcf5b1e61dd3d8fe