Restore dock home support for car docks.
While it's still not generally appropriate for the home key
to do anything other than take you home, a car is a somewhat
different context: because of the user's reduced ability to
interact with the device, it's especially useful to
temporarily replace the launcher with a car-optimized
variant. While it would be ideal for the entire system
(particularly the user's everyday Launcher) to dynamically
reconfigure their UIs upon entering car mode, it's not the
case today, and car home apps are a useful compromise.

Quick test:

	$ adb shell am broadcast \
		-a android.intent.action.DOCK_EVENT \
		--ei android.intent.extra.DOCK_STATE 2

This reverts commit 01386f7e1991809ebd38c4c6494c98356ab22fb6 [formerly 7c135205f2f4c192299ad0215276e9e481bccaa5].

Bug: 8234767
Change-Id: I17588fff46e7f5e3773d3ad202601a5642943444

Former-commit-id: 6396c7205a7c5b16dae5b2e0ba11079005ad7762